Rename file to thaliTape.js
