+++ test/www/jxcore/bv_tests/disabled/testConnectionTable.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var tape = require('../lib/thali-tape');
+var tape = require('../lib/thaliTape');
 var ConnectionTable = require('thali/identityExchange/connectionTable');
 var ThaliReplicationManager = require('thali/thalireplicationmanager');
 var EventEmitter = require('events').EventEmitter;
