+++ test/www/jxcore/bv_tests/disabled/testLargerHashStateMachine.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var tape = require('../lib/thali-tape');
+var tape = require('../lib/thaliTape');
 var LargerHashStateMachine = require('thali/identityExchange/LargerHashStateMachine');
 var identityExchangeUtils = require('thali/identityExchange/identityExchangeUtils');
 var identityExchangeTestUtils = require('./identityExchangeTestUtils');
