+++ test/www/jxcore/bv_tests/testCreatePeerListener.js
@@ -4,7 +4,7 @@ var originalMobile = typeof Mobile === 'undefined' ? undefined : Mobile;
 var mockMobile = require('../lib/MockMobile');
 var net = require('net');
 var multiplex = require('multiplex');
-var tape = require('../lib/thali-tape');
+var tape = require('../lib/thaliTape');
 var ThaliTCPServersManager = require('thali/NextGeneration/mux/thaliTcpServersManager');
 var makeIntoCloseAllServer = require('thali/NextGeneration/makeIntoCloseAllServer');
 var Promise = require('lie');
