+++ test/www/jxcore/bv_tests/testTests.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var tape = require('../lib/thali-tape');
+var tape = require('../lib/thaliTape');
 
 var customData = 'custom data';
