+++ test/www/jxcore/bv_tests/testThaliMobileNative.js
@@ -2,7 +2,7 @@
 
 var net = require('net');
 var randomstring = require('randomstring');
-var tape = require('../lib/thali-tape');
+var tape = require('../lib/thaliTape');
 var makeIntoCloseAllServer = require('thali/NextGeneration/makeIntoCloseAllServer');
 var logger = require('thali/thalilogger')('testThaliMobileNative');
 var Promise = require('lie');
