+++ test/www/jxcore/bv_tests/testThaliPeerDictionary.js
@@ -1,5 +1,5 @@
 'use strict';
-var tape = require('../lib/thali-tape');
+var tape = require('../lib/thaliTape');
 var crypto = require('crypto');
 var sinon = require('sinon');
 var assert = require('assert');
