+++ test/www/jxcore/bv_tests/testThaliPeerPoolInterface.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var tape = require('../lib/thali-tape');
+var tape = require('../lib/thaliTape');
 var PeerAction = require('thali/NextGeneration/thaliPeerPool/thaliPeerAction');
 var connectionTypes =
   require('thali/NextGeneration/thaliMobile').connectionTypes;
