+++ test/www/jxcore/bv_tests/testThaliReplicationUtilities.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var tape = require('../lib/thali-tape');
+var tape = require('../lib/thaliTape');
 var compareBufferArrays =
   require('thali/NextGeneration/replication/utilities').compareBufferArrays;
 var RefreshTimerManager =
