+++ test/www/jxcore/lib/testUtils.js
