+++ test/www/jxcore/meta_tests/testPerfTestFramework.js
@@ -2,7 +2,7 @@
 
 var PerfTestFramework = require('../../../TestServer/PerfTestFramework.js');
 var TestDevice = require('../../../TestServer/TestDevice.js');
-var tape = require('../lib/thali-tape');
+var tape = require('../lib/thaliTape');
 
 var test = tape({
   setup: function (t) {
