+++ thali/install/postPublishThaliCordovaPlugin.js
